packages/ai_agents/pyproject.toml
[project]
requires-python = ">=3.12"
dynamic = ["version"]
name = "what-to-cook"
dependencies = [
    "python-dotenv==1.1.1",
    "strands-agents==1.9.0",
    "strands-agents-tools==0.2.8",
    "strands-agents-builder==0.1.10",
    "bedrock-agentcore==0.1.4",
    "bedrock-agentcore-starter-toolkit==0.1.12",
    "uvicorn[standard]==0.35.0",
]
